1012 중간 commit
initDb.js, src/index.js : 미완성 -> 진도를 못 따라가 가서 추후에 질문

src/route/users.js : 유저 생성 API 작동 시 Permission 테이블에 데이터가 들어가도록 하는 코드 추가와 POST로 넘어오는 데이터에 "permission": {
        "title": "관리자",
        "level": 0
    }
추가

+++ src/index.js
@@ -6,19 +6,35 @@ import db from './models/index.js';
 
 const app = express();
 
-db.sequelize.query("SET FOREIGN_KEY_CHECKS = 0", { raw: true })
-    .then(() => {
-        db.sequelize.sync({ force: true }).then(() => {
-            console.log("sync 끝");
+if (process.env.NODE_ENV === "development") {
+    db.sequelize.query("SET FOREIGN_KEY_CHECKS = 0", { raw: true })
+        .then(() => {
+            db.sequelize.sync({ force: true }).then(() => {
+                console.log("개발환경 sync 끝");
 
-            app.use(express.json());
+                app.use(express.json());
 
-            app.use(express.urlencoded({ extended: true }));
+                app.use(express.urlencoded({ extended: true }));
 
-            app.use("/users", userRouter);
+                app.use("/users", userRouter);
 
-            app.use("/boards", boardRouter);
+                app.use("/boards", boardRouter);
 
-            app.listen(3000);
+                app.listen(3000);
+            });
         });
+} else if (process.env.NODE_ENV === "production") {
+    db.sequelize.sync().then(() => {
+        console.log("사용환경 sync 끝");
+
+        app.use(express.json());
+
+        app.use(express.urlencoded({ extended: true }));
+
+        app.use("/users", userRouter);
+
+        app.use("/boards", boardRouter);
+
+        app.listen(3000);
     });
+}
